[FRONT/BACK] add admin user CRUD (#33)

+++ back/src/Application/User/Handler/DeleteUserCommandHandler.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Application\User\Handler;
 
 use App\Application\User\Command\DeleteUserCommand;
+use App\Domain\Common\Exception\ForbiddenException;
 use App\Domain\Common\Exception\NotFoundException;
 use App\Domain\User\Repository\UserRepositoryInterface;
 use Symfony\Component\Uid\Ulid;
@@ -17,11 +18,14 @@ class DeleteUserCommandHandler
     }
 
     /**
-     * @throws NotFoundException When exchange is not found
+     * @throws NotFoundException When user is not found
      */
     public function __invoke(DeleteUserCommand $command): Ulid
     {
         $user = $this->userRepository->getOneByUid($command->userUid);
+        if (null !== $user->getDeletedAt()) {
+            throw new ForbiddenException(sprintf('User %s is already deleted', $user->getEmail()));
+        }
         $this->userRepository->delete($user);
 
         return $user->getUid();
